Migrate to global Zustand store
We are evaluting how to store shared application state and decided to
use Zustand with a single global store via React context.

This commit moves the existing store for query state to a global one,
but keeps the same API.
