+++ client/web/src/nav/GlobalNavbar.test.tsx
@@ -12,6 +12,7 @@ import {
 import { extensionsController, NOOP_SETTINGS_CASCADE } from '@sourcegraph/shared/src/util/searchTestHelpers'
 
 import { SearchPatternType } from '../graphql-operations'
+import { GlobalStoreProvider } from '../stores/global'
 import { ThemePreference } from '../theme'
 
 import { GlobalNavbar } from './GlobalNavbar'
@@ -68,18 +69,22 @@ describe('GlobalNavbar', () => {
 
     test('default', () => {
         const { asFragment } = render(
-            <MemoryRouter>
-                <GlobalNavbar {...PROPS} />
-            </MemoryRouter>
+            <GlobalStoreProvider>
+                <MemoryRouter>
+                    <GlobalNavbar {...PROPS} />
+                </MemoryRouter>
+            </GlobalStoreProvider>
         )
         expect(asFragment()).toMatchSnapshot()
     })
 
     test('low-profile', () => {
         const { asFragment } = render(
-            <MemoryRouter>
-                <GlobalNavbar {...PROPS} variant="low-profile" />
-            </MemoryRouter>
+            <GlobalStoreProvider>
+                <MemoryRouter>
+                    <GlobalNavbar {...PROPS} variant="low-profile" />
+                </MemoryRouter>
+            </GlobalStoreProvider>
         )
         expect(asFragment()).toMatchSnapshot()
     })
